show all popups before proceeding to the next tutorial proof
