+++ src/components/TestQuestions.tsx
@@ -16,7 +16,7 @@ interface QuestionsProps {
   scaffolding: { [key: string]: boolean };
   updateScaffolding: (questionType: string) => void;
   setActiveQuestionIndex: (index: number) => void;
-  incrementTutorial?: () => void;
+  incrementTutorial?: () => boolean;
 }
 
 export class TestQuestions extends React.Component<QuestionsProps> {
@@ -52,9 +52,16 @@ export class TestQuestions extends React.Component<QuestionsProps> {
       this.props.onNext(1);
     } else if (this.props.questionIdx < this.props.questions.length - 1) {
       this.props.setActiveQuestionIndex(this.props.questionIdx + 1);
-      this.props.incrementTutorial && this.props.incrementTutorial();
+      const shouldUpdate =
+        this.props.incrementTutorial && this.props.incrementTutorial();
     } else {
-      this.props.incrementTutorial && this.props.incrementTutorial();
+      let shouldUpdate = true;
+      if (this.props.incrementTutorial) {
+        shouldUpdate = this.props.incrementTutorial();
+        if (!shouldUpdate) {
+          return;
+        }
+      }
       this.props.onNext(1);
     }
   };
